Version 2 - lets go small

+++ src/mr.roboto/src/mr/roboto/views/runhooks.py
@@ -1,30 +1,15 @@
 # -*- encoding: utf-8 -*-
 from cornice import Service
 from mr.roboto.security import validatetoken
-from mr.roboto.security import validategithub
-from mr.roboto.jenkinsutil import jenkins_core_job
-from mr.roboto.jenkinsutil import jenkins_create_pull_job
-from mr.roboto.jenkinsutil import jenkins_get_job_url
-from mr.roboto.jenkinsutil import jenkins_build_job
-from mr.roboto.jenkinsutil import jenkins_remove_job
 from mr.roboto.buildout import PloneCoreBuildout
-from mr.roboto.subscriber import get_info_from_commit
-
-from mr.roboto.db import CorePackages
-from mr.roboto.db import CorePackage
-from mr.roboto.db import JenkinsJob
-from mr.roboto.db import JenkinsJobs
-from mr.roboto.db import PullRequest
-from mr.roboto.db import PullRequests
-
-import transaction
 
 import logging
+import pickle
 import json
-import uuid
 
-
+debug = False
 logger = logging.getLogger('mr.roboto')
+
 
 def add_log(request, who, message):
     logger.info(who + " " + message)
@@ -36,26 +21,24 @@ createGithubPostCommitHooks = Service(
 )
 
 
-@createGithubPostCommitHooks.post()
+@createGithubPostCommitHooks.get()
 @validatetoken
 def createGithubPostCommitHooksView(request):
+    # sources_dict
+    # {('package_path', 'branch'): ['5.0', '4.3']}
+    #
+    # checkouts_dict
+    # {'5.0': ['package', '...']}
+
+    sources_dict = {}
+    checkouts_dict = {}
     # We should remove all the actual hooks
     github = request.registry.settings['github']
     roboto_url = request.registry.settings['roboto_url']
+    sources_file = request.registry.settings['sources_file']
+    checkouts_file = request.registry.settings['checkouts_file']
     actual_plone_versions = request.registry.settings['plone_versions']
 
-    dm = request.registry.settings['dm']
-    core_packages = CorePackages(dm)
-
-    # Remove all core_packages
-    request.registry.settings['db']['core_package'].remove({})
-    # actual_packages = core_packages.keys()
-    # for actual_package in actual_packages:
-    #     del core_packages[actual_package]
-
-    transaction.commit()
-
-    core_packages = CorePackages(dm)
     # Clean Core Packages DB and sync to GH
     for plone_version in actual_plone_versions:
 
@@ -65,9 +48,23 @@ def createGithubPostCommitHooksView(request):
         for source in sources:
             source_obj = buildout.sources[source]
             if source_obj.path is not None:
-                core_packages[source] = CorePackage(source, source_obj.path, source_obj.branch, plone_version)
+                if (source_obj.path, source_obj.branch) not in sources_dict:
+                    sources_dict[(source_obj.path, source_obj.branch)] = [plone_version]
+                else:
+                    sources_dict[(source_obj.path, source_obj.branch)].append(plone_version)
 
-        transaction.commit()
+        checkouts_dict[plone_version] = []
+        for checkout in buildout.checkouts.data:
+            if checkout != '':
+                checkouts_dict[plone_version].append(checkout)
+
+    sf = open(sources_file, 'w')
+    sf.write(pickle.dumps(sources_dict))
+    sf.close()
+
+    sf = open(checkouts_file, 'w')
+    sf.write(pickle.dumps(checkouts_dict))
+    sf.close()
 
     # hooks URL
     commit_url = roboto_url + 'run/corecommit'
@@ -85,14 +82,20 @@ def createGithubPostCommitHooksView(request):
             #if hook.name == 'web' and (hook.config['url'].find(roboto_url) or hook.config['url'].find('jenkins.plone.org')):
             if hook.name == 'web' and hook.config['url'].find(roboto_url):
                 add_log(request, 'github', 'Removing hook ' + str(hook.config))
-                hook.delete()
+                if debug:
+                    print "Debug removing hook"
+                else:
+                    hook.delete()
 
         # Add the new hooks
         add_log(request, 'github', 'Creating hook ' + commit_url + ' and ' + pull_url)
         messages.append('Creating hook ' + commit_url)
         try:
-            repo.create_hook('web', {'url': commit_url, 'secret': request.registry.settings['api_key']}, 'push', True)
-            repo.create_hook('web', {'url': pull_url, 'secret': request.registry.settings['api_key']}, 'pull_request', True)
+            if debug:
+                print "Debug creating hook"
+            else:
+                repo.create_hook('web', {'url': commit_url, 'secret': request.registry.settings['api_key']}, 'push', True)
+                repo.create_hook('web', {'url': pull_url, 'secret': request.registry.settings['api_key']}, 'pull_request', True)
         except:
             pass
     return json.dumps(messages)
